requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
langchain
langgraph
requests
python-dotenv
httpx
groq
pandas
numpy
scikit-learn
streamlit
apscheduler
python-jose[cryptography]==3.3.0
passlib[bcrypt]
python-multipart
bcrypt
agentlightning
google-genai
pulp
scipy
pydantic-settings
langchain-community
langchain-groq